requirements.txt
asgiref==3.6.0
coverage==7.2.3
decouple==0.0.7
Django==4.1.4
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-redis==5.4.0
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15

